Add aliases for elf_getshnum and elf_getshstrndx.
Sun in their implementation messed up the semantics of these functions.
To get back to source code compatibility they asked for a solution.
We simply add aliases for the two affected functions, named
elf_getshdrnum and elf_getshdrstrndx.  The semantics remains the
same.  Users of the old interfaces receive a warning for now.  In
future we might make it impossible to link programs which request
the old interfaces (binary compatibility will of course be maintained).