tests: backtrace.c only check we caught the last instruction on x86_64.
On some architectures gcc might introduce some "padding instructions"
at the end of the function (like on ppc64). So only assert we are at the
last instruction of backtracegen if on x86_64 native. In theory the assert
could even fail on that architectures, but in practice it doesn't and it
is a nice test to have.

Signed-off-by: Mark Wielaard <mjw@redhat.com>