readelf: Read the decompressed data when processing .zdebug DWARF sections.
readelf uses libdw to open the Dwarf and read some of the DWARF data.
But it also uses its own parsers to display some of the low-level
unprocessed data. If the DWARF debug section was zlib compressed it
should actually use the decompressed section data from libdw instead
of the raw section data.

Includes a testcase for those sections that couldn't be properly
displayed when compressed before.

Signed-off-by: Mark Wielaard <mjw@redhat.com>