libelf: Make sure conversion functions work on aligned data for type.
The gelf_xlate conversion functions work on properly aligned ELF data
types.  If elf_get data needs to do conversion and ! ALLOW_UNALIGNED
and the rawdata_base isn't aligned properly for the section type, then
provide an aligned copy of the data.

Found with --enable-sanitize-undefined  in run-test-archive64.sh on x86_64.

Signed-off-by: Mark Wielaard <mjw@redhat.com>