config: Use ARFLAGS 'cr'.
By default automake uses ARFLAGS cru. 'u' is an optimization that
makes sure files are only added to the archive when they are newer
than the version already in the .a file. This optimization doesn't
work when ar is deterministic by default because then it doesn't
record timestamps in the archive. Deterministic ar is the default
now on various distributions causing multiple warnings like:

  ar: `u' modifier ignored since `D' is the default (see `U')

To suppress these warnings drop the 'u' flag.

Signed-off-by: Mark Wielaard <mjw@redhat.com>